fix(build): clang -Werror portability — main was red on ninja-clang/macOS
The ninja-clang (Linux) and macOS clang CI jobs build with
-Werror -Wconversion -Wshadow -Wpedantic -Wmissing-field-initializers,
which MSVC /W4 did not flag. Two committed changes broke those jobs:

- data_dict.cpp: IndexEntry grew per-tag metadata fields (865dfa3b), so
  the 3-element brace-init `{alias, path, {}}` tripped
  -Werror=missing-field-initializers. Name the object and assign instead.
- dd_native.cpp (parked OADD codec, 0a809b2a): the `<<8` in u16()
  integer-promotes its operands to int, so the OR result is int and the
  assignment narrows — -Werror=implicit-int-conversion. Cast the whole
  expression back to uint16_t. (u32() is exempt: uint32_t outranks int.)

Verified locally by reproducing the ninja-clang preset build + full ctest
(100% pass) and the msvc-x86 build.

+++ src/engine/dd_native.cpp
@@ -43,8 +43,13 @@ struct Reader {
     }
     std::uint16_t u16() {
         if (!need(2)) return 0;
-        std::uint16_t v = static_cast<std::uint16_t>(p[pos]) |
-                          (static_cast<std::uint16_t>(p[pos + 1]) << 8);
+        // RCB 07/16/2026: the <<8 integer-promotes the operands to int, so the
+        // whole OR is int — cast back explicitly or clang -Werror=conversion
+        // flags the narrowing on assignment (u32 below is exempt: uint32_t
+        // outranks int, so no promotion occurs).
+        std::uint16_t v = static_cast<std::uint16_t>(
+                          static_cast<std::uint16_t>(p[pos]) |
+                          (static_cast<std::uint16_t>(p[pos + 1]) << 8));
         pos += 2; return v;
     }
     std::uint32_t u32() {
